+++ param_global/hooks.py
@@ -37,6 +37,8 @@ app_include_js = [
 	"pg_link_search.bundle.js",
 	"pg_nombres.bundle.js",
 	"pg_verrou.bundle.js",
+	"pg_grille_entree.bundle.js",
+	"pg_tri_dropdown.bundle.js",
 	"pg_environnement.bundle.js",
 ]